Remove nvidia x11 drivers from @hardware-support group
When the optional packages of @hardware-support are requested, a set
of xorg-x11-nvidia* drivers get pulled in. Some of them (e.g. 304xx)
drop in their own /etc/X11/*xorg.conf file, assuming the relvant
hardware is present, and, if that's not the case, breaking X11 hardware
autodetection in the process.

Signed-off-by: Gabriel Somlo <gsomlo@gmail.com>